v2/requirements.txt
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
requests==2.31.0
python-dotenv==1.0.1
tavily-python==0.5.0
wikipedia 
pydantic